fix: work around msvc compiler bugs related to NTTP/array refs in param packs
It's unclear what exactly is going on here. But use of a reference to
bounded-array as a template parameter inside a parameter pack cause MSVC
to emite useless compiler errors. On MSVC 19.16 it appears to eliminate
these overloads from consideration without any diagnostic (like SFINAE).
On more recent versions it provides the supremely useless "failed to
specialize function template" without informing about *why* it is
failing to do so.

And some variations of these, applying decltype to NTTPs, even ICEs the
compiler.

Converting `const T (&)[N]` to `array_ref<T, N>` allows us to *not* have
a reference-to-array expression inside the parameter pack and this
appears to stop triggering all compiler bugs discovered thus far.